+++ start-master.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
 NAME="spark-master"
-HOST_IP=$(ip route get 8.8.8.8 | awk '{print $NF; exit}')
+HOST_IP=$1
+if [ "$HOST_IP" == "" ]; then
+	HOST_IP=$(ip route get 8.8.8.8 | awk '{print $NF; exit}')
+fi
+
+echo "MASTER IP: "$HOST_IP
+
 docker stop $NAME
 docker rm $NAME
 __image="jupyter/all-spark-notebook"
@@ -22,5 +28,3 @@ docker exec -d $NAME $__path_to_spark/sbin/start-master.sh
 echo "The spark ui can be seen in a web browser at: http://"$HOST_IP":8080"
 echo "run the worker on each worker node via..."
 echo "bash start-worker.sh spark://"$HOST_IP":7077"
-
-
